requirements.txt
-i https://pypi.org/simple

tensorflow-gpu = "==1.12.0"
ipython = "*"
numpy = "*"
scipy = "*"
pillow = "*"
cython = "*"
matplotlib = "*"
opencv-python = "*"
h5py = "*"
imgaug = "*"
opencv-contrib-python = "*"
pandas = "*"
scikit-learn = "*"
tensorboard = "==1.6"
keras = "==2.0.8"
imutils = "*"
